feat: Implement static analysis of the class manually
In the beginning, we use tsc to compile the whole file before other compiler,
and so that metadata can be injected to class member,
but it is destructive, now we decorate the member by traversing the whole ast.
But it might be unstable, we should me test cases. welcome PR.